+++ packages/expo-router/src/renderRootComponent.tsx
@@ -1,8 +1,7 @@
 import { registerRootComponent } from 'expo';
+import * as SplashScreen from 'expo-splash-screen';
 import React from 'react';
 import { Platform, View } from 'react-native';
-
-import { SplashScreen, _internal_preventAutoHideAsync } from './views/Splash';
 
 function isBaseObject(obj: any) {
   if (Object.prototype.toString.call(obj) !== '[object Object]') {
@@ -54,7 +53,8 @@ export function renderRootComponent(Component: React.ComponentType<any>) {
   try {
     // This must be delayed so the user has a chance to call it first.
     setTimeout(() => {
-      _internal_preventAutoHideAsync();
+      // @ts-expect-error: This function is native-only and for internal-use only.
+      SplashScreen._internal_preventAutoHideAsync?.();
     });
 
     if (process.env.NODE_ENV !== 'production') {